ege_path模块分离，编译体积预计降10K
最小程序：455K->443.5K
